+++ docs/.vitepress/config.mjs
@@ -13,7 +13,7 @@ export default defineConfig({
             {text: 'Home', link: '/'},
             {text: 'Guide', link: '/guide/'},
             //{text: 'Reference', link: '/docs/'},
-            //{text: 'Course', link: '/course/'}
+            {text: 'Course', link: '/course/'}
         ],
 
         footer: {
@@ -78,6 +78,45 @@ export default defineConfig({
                 //},
                 {text: 'Resources', link: '/guide/resources'},
                 {text: 'Community', link: '/guide/community'}
+            ],
+            '/course/': [
+                {
+                    text: 'Video Course',
+                    items: [
+                        {
+                            text: 'Introduction',
+                            items: [
+                                { text: 'Welcome (3:35)', link: '/course/video-01' },
+                            ]
+                        },
+                        {
+                            text: 'Getting Set Up',
+                            items: [
+                                { text: 'Installation (5:48)', link: '/course/video-02' },
+                                { text: 'Core Setup (7:24)', link: '/course/video-03' },
+                                { text: 'Scaffolding an App (12:34)', link: '/course/video-04' },
+                                { text: 'Enabling OData (8:52)', link: '/course/video-05' },
+                                { text: 'Configuration (7:52)', link: '/course/video-06' },
+                            ]
+                        },
+                        {
+                            text: 'Building the UI',
+                            items: [
+                                { text: 'The Master View (15:08)', link: '/course/video-07' },
+                                { text: 'Live Search (17:16)', link: '/course/video-08' },
+                                { text: 'The Detail Page (13:16)', link: '/course/video-09' },
+                                { text: 'SAP & the Rest of Us (6:53)', link: '/course/video-10' },
+                                { text: 'The Lock Function (24:44)', link: '/course/video-11' },
+                            ]
+                        },
+                        {
+                            text: 'Wrap-Up',
+                            items: [
+                                { text: 'Just the Beginning (10:20)', link: '/course/video-12' },
+                            ]
+                        }
+                    ],
+                }
             ]
         },